+++ joyyios/JYYRS.h
@@ -9,14 +9,14 @@
 @interface JYYRS : NSObject
 
 + (instancetype)yrsWithYob:(NSUInteger)yob region:(NSUInteger)region sex:(NSUInteger)sex;
-+ (instancetype)yrsWithValue:(NSUInteger)value;
++ (instancetype)yrsWithValue:(uint64_t)value;
 
 - (instancetype)initWithYob:(NSUInteger)yob region:(NSUInteger)region sex:(NSUInteger)sex;
-- (instancetype)initWithValue:(NSUInteger)value;
+- (instancetype)initWithValue:(uint64_t)value;
 
 @property (nonatomic) NSUInteger yob;
 @property (nonatomic) NSUInteger region;
 @property (nonatomic) NSUInteger sex;
-@property (nonatomic) NSUInteger value;
+@property (nonatomic) uint64_t value;
 
 @end
